pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
alembic = "1.13.1"
assemblyai = "0.22.0"
boto3 = "1.34.47"
botocore = "1.34.47"
celery = "5.3.6"
Flask = "3.0.2"
Flask-JWT-Extended = "4.6.0"
flask-marshmallow = "1.2.0"
Flask-SQLAlchemy = "3.1.1"
gunicorn = "21.2.0"
redis = "5.0.1"
requests = "2.31.0"
Werkzeug = { version = "3.0.1", extras=["watchdog"] }
python-dotenv = "^1.0.1"
marshmallow-sqlalchemy = "^1.0.0"
openai = "^1.14.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
